created setAfterBuilDTicks function
